Move ThemeComponents to src/components
