+++ src/components/Calendar/Events.jsx
@@ -3,7 +3,7 @@ import "utility/fonts.css";
 import SubpageOuter from "components/SubpageOuter/SubpageOuter.jsx";
 import Calendar from "./Calendar.jsx";
 import Navbar from "components/Navbar.jsx";
-import { SubTheme } from "ThemeComponents.jsx";
+import { SubTheme } from "components/ThemeComponents.jsx";
 
 const CalendarStyle = {
     color: "#F15D24",
