+++ src/components/Navbar.jsx
@@ -5,7 +5,7 @@ import devices from "../utility/MediaQueries.js";
 import logoUrl from "assets/logo-orange-cropped.png";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { SOCIAL_MEDIA_LINKS } from "utility/constants.js";
-import { SubTheme } from "ThemeComponents.jsx";
+import { SubTheme } from "components/ThemeComponents.jsx";
 
 const FlexWrapper = styled.div`
     height: 80px;
